pyproject.toml
[tool.poetry]
name = "rssfeeder"
version = "0.1.0"
description = ""
authors = ["Jakub Dvořák <dvorak1430@seznam.cz>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
feedparser = "^6.0.8"
colorama = "^0.4.4"
httpx = "^0.21.3"
snakeviz = "^2.1.1"
rich = "^11.0.0"
install = "^1.3.5"
curser = "^1.0.1"
pynput = "^1.7.6"
xlib = "^0.21"
tk = "^0.1.0"
TKinterModernThemes = "^1.9.3"
flake8 = "^6.0.0"
pyright = "^1.1.310"
black = "^23.3.0"
async-tkinter-loop = "^0.8.1"
Pillow = "^9.5.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
